fixed some data types

+++ src/main/java/net/pentlock/thunderdataengine/utilities/HouseUtil.java
@@ -27,7 +27,7 @@ public class HouseUtil {
      * @return this is the house object after it has been created
      */
 
-    public static House createHouse(String name, double cost, double unitsAvailable, double unitsIssued, String[] doorLocation, String[] arrivalLocation,
+    public static House createHouse(String name, double cost, int unitsAvailable, int unitsIssued, String[] doorLocation, String[] arrivalLocation,
                                     Map<UUID, String[]> instances, Map<UUID, String[]> instanceLocation) {
 
         House house = new House(name, cost,unitsAvailable, unitsIssued, doorLocation, arrivalLocation, instances, instanceLocation);
